requirements.txt
Django>=1.8.10
uwsgi
gunicorn==19.4.5
django-filter
djangorestframework>=3.5
MySQL-python
Pillow
redis
bson
pymongo
pysnmp
xlwt
xlrd
python-consul
GRR-M2Crypto
django-cors-headers
dulwich
funky
celery==3.1.18
django-celery
celery-with-redis
kombu==3.0.37
python-jenkins
channels